🔍 refactor.colorIcons element 정리
