+++ src/components/modal/optionModal/OptionModal.tsx
@@ -5,7 +5,7 @@ import { openDB } from "idb";
 
 import { MainButton } from "../../../elements/buttons/Buttons";
 import Modal from "../Modal";
-import { ColorHoverIcon } from "../../../elements/ColorIcons";
+import { ColorHoverIcon } from "../../../elements/colorIcons/ColorIcons";
 import { useGetOption } from "./useGetOption";
 // import ProdCnt, { ProdOptCnt } from "../../detailPage/prodCnt/ProdCnt";
 export interface ModalType {
