+++ src/elements/colorIcons/colorIcons.type.ts
@@ -1,0 +1,5 @@
+export interface ColorIconType {
+  colorCode?: string | number;
+  colorName?: string;
+  onClick?: () => void;
+}
